put TokenInfo into seperate file
